requirements.txt
streamlit==1.35.0
pydantic==2.7.1
huggingface_hub==0.23.2
python-dotenv==1.0.1